daemon: Avoid comparison of integers of different signs.
Fixes guix/guix#1173.

* nix/libstore/gc.cc (LocalStore::removeUnusedLinks): Cast ‘st_size’ as
unsigned.

Reported-by: Congcong Kuo <congcong.kuo@gmail.com>

+++ nix/libstore/gc.cc
@@ -608,7 +608,8 @@ void LocalStore::removeUnusedLinks(const GCState & state)
 
 	/* Drop links for files smaller than 'deduplicationMinSize', even if
 	   they have more than one hard link.  */
-        if (st.st_nlink != 1 && st.st_size >= deduplicationMinSize) {
+        if (st.st_nlink != 1
+	    && ((unsigned long long) st.st_size) >= deduplicationMinSize) {
             actualSize += st.st_size;
             unsharedSize += (st.st_nlink - 1) * st.st_size;
             continue;
